Basic Types

Signature

ColorProfiles.Signature — Type

Signature{id}

Signature is a sigleton type to represent an 4-ASCII-character symbol. The id paramater should be a UInt32 number.

Note

Note that the semantics of signatures are context dependent and are not globally unique in profiles. For example, "XYZ "(0x58595a20) is used for ColorSpaces.XYZSpace and TaggedTypes.TaggedXYZ.

Basic Number Types

ColorProfiles.Position — Type

ColorProfiles.Position

Indicate positions of some data elements.

ColorProfiles.Response16 — Type

ColorProfiles.Response16

Associate a normalized device code with a measurement value.

ColorProfiles.UQ16f16 — Type

ColorProfiles.UQ16f16

A fixed unsigned 32-bit number with 16 fractional bits.

Compat

This type may be implemented by FixedPointNumbers in the feature.

ColorProfiles.UQ1f15 — Type

ColorProfiles.UQ1f15

A fixed unsigned 16-bit number with 15 fractional bits.

Compat

This type may be implemented by FixedPointNumbers in the feature.

ColorProfiles.UQ8f8 — Type

ColorProfiles.UQ8f8

A fixed unsigned 16-bit number with 8 fractional bits.

Compat

This type may be implemented by FixedPointNumbers in the feature.

Enumeration types and constants

CMM types

ColorProfiles.CMMTypes — Module

CMMTypes module defines a type for identifying CMM (Color Management Module) types. See CMMType.

ColorProfiles.CMMTypes.CMMType — Type

CMMType{sig}

A singleton type for CMM (Color Management Module) types. The sig parameter should be a Signature instance.

Currently, the CMMTypes module does not define the CMMType instances (constants). See Signature Registry in the ICC website for the registered signatures. For example, you can use CMMType(:lcms) for the Little CMS CMM, as a singleton instance.

Profile/device classes

ColorProfiles.ProfileClasses — Module

ProfileClasses module defines the type for identifying profile/device classes and its instances. See ProfileClass.

ColorProfiles.ProfileClasses.ProfileClass — Type

ProfileClass{sig}

InputClass

DisplayClass

OutputClass

DeviceLinkClass

ColorSpaceClass

AbstractClass

NamedColorClass

ColorProfiles.ProfileClasses.AbstractClass — Constant

AbstractClass

ColorProfiles.ProfileClasses.ColorSpaceClass — Constant

ColorSpaceClass

ColorProfiles.ProfileClasses.DeviceLinkClass — Constant

DeviceLinkClass

ColorProfiles.ProfileClasses.DisplayClass — Constant

DisplayClass

A singleton instance of ProfileClass{Signature(:mntr)}. This is one of basic device classes.

ColorProfiles.ProfileClasses.InputClass — Constant

InputClass

A singleton instance of ProfileClass{Signature(:scnr)}. This is one of basic device classes.

ColorProfiles.ProfileClasses.NamedColorClass — Constant

NamedColorClass

ColorProfiles.ProfileClasses.OutputClass — Constant

OutputClass

A singleton instance of ProfileClass{Signature(:prtr)}. This is one of basic device classes.

Color Spaces

ColorProfiles.ColorSpaces — Module

ColorSpaces module defines a type for identifying color spaces, and its instances (constants). See ColorSpace.

ColorProfiles.ColorSpaces.ColorSpace — Type

ColorSpace{sig}

ColorSpace is a singleton type to represent the data color space. The sig parameter should be a Signature instance.

The following singleton instances are defined.

Color Space	Signature	Color Space	Signature

XYZSpace	"XYZ "	Color2Space	"2CLR"

LabSpace	"Lab "	Color3Space	"3CLR"

LuvSpace	"Luv "	Color4Space	"4CLR"

YCbCrSpace	"YCbr"	Color5Space	"5CLR"

YxySpace	"Yxy "	Color6Space	"6CLR"

RGBSpace	"RGB "	Color7Space	"7CLR"

GraySpace	"GRAY"	Color8Space	"8CLR"

HSVSpace	"HSV "	Color9Space	"9CLR"

HLSSpace	"HLS "	Color10Space	"ACLR"

CMYKSpace	"CMYK"	Color11Space	"BCLR"

CMYSpace	"CMY "	Color12Space	"CCLR"

Color13Space	"DCLR"

Color14Space	"ECLR"

Color15Space	"FCLR"

ColorProfiles.ColorSpaces.LabSpace — Constant

LabSpace refers to CIELAB or PCSLAB color space depending upon the context.

ColorProfiles.ColorSpaces.XYZSpace — Constant

XYZSpace refers to nCIEXYZ or PCSXYZ color space depending upon the context.

Platforms

ColorProfiles.Platforms — Module

Platforms module defines a type for

ColorProfiles.Platforms.Platform — Type

Platform{sig}

Platform is a singleton type to represent the primary platform. The sig parameter should be a Signature instance.

The Platforms module defines the following singleton instances, but they are not exported:

Instance Name	Primary Platform	Signature

ApplePlatform	Apple Computer, Inc.	"APPL"

MicrosoftPlatform	Microsoft Corporation	"MSFT"

SGIPlatform	Silicon Graphics, Inc.	"SGI "

SunPlatform	Sun Microsystems, Inc.	"SUNW"

TaligentPlatform	Taligent, Inc.	"TGNT"

Manufacturers

ColorProfiles.Manufacturers — Module

Manufacturers module defines a type for identifying the manufacturers or creators. See Manufacturer.

ColorProfiles.Manufacturers.Manufacturer — Type

Manufacturer{sig}

Currently, the Manufacturers module does not define the Manufacturer instances (constants). See Manufacturer Registry in the ICC website.

Device models

ColorProfiles.DeviceModels — Module

DeviceModels module defines a type for identifying the device model. See DeviceModel.

ColorProfiles.DeviceModels.DeviceModel — Type

DeviceModel{sig}

Currently, the DeviceModels module does not define the DeviceModel instances (constants). See Device Registry in the ICC website.

Rendering intents

ColorProfiles.RenderingIntents — Module

RenderingIntents module defines the type for identifying rendering intents and its instances (constants). See RenderingIntent.

ColorProfiles.RenderingIntents.RenderingIntent — Type

RenderingIntent{id}

Rendering Intent	Encoded id

Perceptual	0x00000001

RelativeColorimetric	0x00000002

Saturation	0x00000003

AbsoluteColorimetric	0x00000004

ColorProfiles.RenderingIntents.AbsoluteColorimetric — Constant

AbsoluteColorimetric

ICC-absolute colorimetric rendering intent.

ColorProfiles.RenderingIntents.Perceptual — Constant

Perceptual

Perceptual rendering intent.

ColorProfiles.RenderingIntents.RelativeColorimetric — Constant

RelativeColorimetric

Media-relative colorimetric rendering intent.

ColorProfiles.RenderingIntents.Saturation — Constant

Saturation

Saturation rendering intent.

Standard Illuminants

ColorProfiles.StandardIlluminants — Module

StandardIlluminants module defines the type for identifying standard illuminants and its instances (constants). See StandardIlluminant.

ColorProfiles.StandardIlluminants.StandardIlluminant — Type

StandardIlluminant{id}

Standard Illuminant	Encoded id

IlluminantUnknown	0x00000000

IlluminantD50	0x00000001

IlluminantD65	0x00000002

IlluminantD93	0x00000003

IlluminantF2	0x00000004

IlluminantD55	0x00000005

IlluminantA	0x00000006

IlluminantE	0x00000007

IlluminantF8	0x00000008

ColorProfiles.StandardIlluminants.IlluminantD50 — Constant

IlluminantD50

ColorProfiles.StandardIlluminants.IlluminantD65 — Constant

IlluminantD65

This is used as the default white point in Colors.

ColorProfiles.StandardIlluminants.IlluminantE — Constant

IlluminantE

Equi-Power standard illuminant.