Tags

ColorProfiles.Tags.AToB0Tag — Type

AToB0Tag{T}

AToB0Tag is the alias for AToBTag{Signature(:A2B0)}.

This tag describes a color transform depending on the device/profile class. Except for AbstractClass and DeviceLinkClass, the transform is mainly for Parceptual rendering.

ProfileClass	from	to

InputClass	Device	PCS

DisplayClass	Device	PCS

OutputClass	Device	PCS

ColorSpaceClass	Color encoding	PCS

AbstractClass	PCS	PCS

DeviceLinkClass	Device1	Device2

ColorProfiles.Tags.AToB1Tag — Type

AToB1Tag{T}

AToB1Tag is the alias for AToBTag{Signature(:A2B1)}.

The transform is mainly for RelativeColorimetric/ AbsoluteColorimetric rendering.

ProfileClass	from	to

InputClass	Device	PCS

DisplayClass	Device	PCS

OutputClass	Device	PCS

ColorSpaceClass	Color encoding	PCS

ColorProfiles.Tags.AToB2Tag — Type

AToB2Tag{T}

AToB2Tag is the alias for AToBTag{Signature(:A2B2)}.

Saturation

ProfileClass	from	to

InputClass	Device	PCS

DisplayClass	Device	PCS

OutputClass	Device	PCS

ColorSpaceClass	Color encoding	PCS

ColorProfiles.Tags.AToBTag — Type

AToBTag{sig, T}

AToBTag defines a color transform to (mainly) PCS. See also BToATag. This is the parametric type for the following tags:

AToB0Tag

AToB1Tag

AToB2Tag

Fields

lut::T: lookup table

ColorProfiles.Tags.BToA0Tag — Type

BToA0Tag{T}

ColorProfiles.Tags.BToA1Tag — Type

BToA1Tag{T}

ColorProfiles.Tags.BToA2Tag — Type

BToA2Tag{T}

ColorProfiles.Tags.BToATag — Type

BToATag{sig, T}

BToATag is This is the parametric type for the following tags:

BToA0Tag

BToA1Tag

BToA2Tag

Fields

lut::T: lookup table

ColorProfiles.Tags.BToD0Tag — Type

BToD0Tag{T}

ColorProfiles.Tags.BToD1Tag — Type

BToD1Tag{T}

ColorProfiles.Tags.BToD2Tag — Type

BToD2Tag{T}

ColorProfiles.Tags.BToD3Tag — Type

BToD3Tag{T}

ColorProfiles.Tags.BToDTag — Type

BToDTag{sig, T}

BToD0Tag

BToD1Tag

BToD2Tag

BToD3Tag

ColorProfiles.Tags.BlueMatrixColumnTag — Type

BlueMatrixColumnTag

ColorProfiles.Tags.BlueTRCTag — Type

BlueTRCTag{T}

ColorProfiles.Tags.CalibrationDateTimeTag — Type

CalibrationDateTimeTag

ColorProfiles.Tags.CharTargetTag — Type

CharTargetTag

ColorProfiles.Tags.ChromaticAdaptationTag — Type

ChromaticAdaptationTag

ColorProfiles.Tags.ChromaticityTag — Type

ChromaticityTag

ColorProfiles.Tags.ColorantOrderTag — Type

ColorantOrderTag

ColorProfiles.Tags.ColorantTableOutTag — Type

ColorantTableOutTag

ColorProfiles.Tags.ColorantTableTag — Type

ColorantTableTag

ColorProfiles.Tags.ColorimetricIntentImageStateTag — Type

ColorimetricIntentImageStateTag

ColorProfiles.Tags.CopyrightTag — Type

CopyrightTag

ColorProfiles.Tags.GamutTag — Type

GamutTag{T}

If the output value of LUT is zero, the PCS color is in-gamut, otherwise, the PCS color is out-of-gamut.

ColorProfiles.Tags.GrayTRCTag — Type

GrayTRCTag{T}

See also [TRCTag](@ref).

ColorProfiles.Tags.GreenTRCTag — Type

GreenTRCTag{T}

See also [TRCTag](@ref), [RedTRCTag](@ref) and [BlueTRCTag](@ref).

ColorProfiles.Tags.LuminanceTag — Type

LuminanceTag = XYZTag{sig"lumi"}

LuminanceTag is the alias for XYZTag{Signature(:lumi)}.

The xyz field is the "absolute" luminance of emissive devices in cd/m^2.

ColorProfiles.Tags.MatrixColumnTag — Type

MatrixColumnTag{sig}

ColorProfiles.Tags.MediaBlackPointTag — Type

MediaBlackPointTag = XYZTag{sig"bkpt"}

(ICC Profile v2)

ColorProfiles.Tags.MediaWhitePointTag — Type

MediaWhitePointTag = XYZTag{sig"wtpt"}

MediaWhitePointTag is the alias for XYZTag{Signature(:wtpt)}.

This tag specifies the chromatically adapted CIEXYZ values of the media white point. The xyz field is in normalized CIEXYZ format (i.e. xyz.y == 1.0).

ColorProfiles.Tags.NamedColor2Tag — Type

NamedColor2Tag

ColorProfiles.Tags.PerceptualRenderingIntentGamutTag — Type

PerceptualRenderingIntentGamutTag

ColorProfiles.Tags.Preview0Tag — Type

Preview0Tag{T} = PreviewTag{sig"pre0",T}

ColorProfiles.Tags.Preview1Tag — Type

Preview1Tag{T} = PreviewTag{sig"pre1",T}

ColorProfiles.Tags.Preview2Tag — Type

Preview2Tag{T} = PreviewTag{sig"pre2",T}

ColorProfiles.Tags.PreviewTag — Type

PreviewTag{sig, T}

ColorProfiles.Tags.ProfileDescriptionTag — Type

ProfileDescriptionTag <: AbstractTag{sig"desc"}

ColorProfiles.Tags.RedTRCTag — Type

RedTRCTag{T} = TRCTag{sig"rTRC",T}

See also TRCTag, GreenTRCTag and BlueTRCTag.

ColorProfiles.Tags.SaturationRenderingIntentGamutTag — Type

SaturationRenderingIntentGamutTag = RenderingIntentGamutTag{sig"rig2"}

ColorProfiles.Tags.TRCTag — Type

TRCTag{sig}

TRC (Tone Reproduction Curve).

RedTRCTag

GreenTRCTag

BlueTRCTag

GrayTRCTag

ColorProfiles.Tags.TechnologyTag — Type

TechnologyTag <: AbstractTag{sig"tech"}

device technology.

Fields

technology::DeviceTechnology

see DeviceTechnology

ColorProfiles.Tags.UnknownTag — Type

UnknownTag{sig, T}

ColorProfiles.Tags.ViewingConditionsTag — Type

ViewingConditionsTag <: AbstractTag{sig"view"}

Fields

illuminant::XYZ{Float64}

surround::XYZ{Float64}

standard::StandardIlluminant

ColorProfiles.Tags.XYZTag — Type

XYZTag{sig}